Refactor

+++ src/base/gcewing/sg/BaseItemBlock.java
@@ -6,16 +6,18 @@
 
 package gcewing.sg;
 
-import net.minecraft.block.*;
-import net.minecraft.item.*;
+import gcewing.sg.BaseMod.IItem;
+import gcewing.sg.BaseMod.ModelSpec;
+import net.minecraft.block.Block;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.ItemBlock;
+import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.world.World;
 
-import gcewing.sg.BaseMod.*;
-
-import static gcewing.sg.BaseBlockUtils.*;
-import static gcewing.sg.BaseUtils.*;
+import static gcewing.sg.BaseBlockUtils.getBlockStateFromMeta;
+import static gcewing.sg.BaseBlockUtils.getMetaFromBlockState;
+import static gcewing.sg.BaseUtils.facings;
 
 public class BaseItemBlock extends ItemBlock implements IItem {
 
@@ -30,7 +32,7 @@ public class BaseItemBlock extends ItemBlock implements IItem {
     public ModelSpec getModelSpec(ItemStack stack) {
         return null;
     }
-    
+
     public int getNumSubtypes() {
         return 1;
     }
@@ -60,35 +62,32 @@ public class BaseItemBlock extends ItemBlock implements IItem {
             return false;
         else if (world.canPlaceEntityOnSide(this.field_150939_a, x, y, z, false, side, player, stack)) {
             int i1 = this.getMetadata(stack.getItemDamage());
-            BaseBlock baseBlock = (BaseBlock)this.field_150939_a;
+            BaseBlock baseBlock = (BaseBlock) this.field_150939_a;
             IBlockState state = baseBlock.onBlockPlaced(world, new BlockPos(x, y, z), facings[side], hitX, hitY, hitZ, i1, player);
             int j1 = baseBlock.getMetaFromState(state);
             if (placeBlockAt(stack, player, world, x, y, z, side, hitX, hitY, hitZ, j1)) {
-                world.playSoundEffect((double)((float)x + 0.5F), (double)((float)y + 0.5F), (double)((float)z + 0.5F), this.field_150939_a.stepSound.func_150496_b(), (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F, this.field_150939_a.stepSound.getPitch() * 0.8F);
+                world.playSoundEffect((float) x + 0.5F, (float) y + 0.5F, (float) z + 0.5F, this.field_150939_a.stepSound.func_150496_b(), (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F, this.field_150939_a.stepSound.getPitch() * 0.8F);
                 --stack.stackSize;
             }
             return true;
-        }
-        else
+        } else
             return false;
     }
 
     @Override
     public boolean placeBlockAt(ItemStack stack, EntityPlayer player, World world, int x, int y, int z,
-        int side, float hitX, float hitY, float hitZ, int meta)
-    {
+                                int side, float hitX, float hitY, float hitZ, int meta) {
         BlockPos pos = new BlockPos(x, y, z);
         IBlockState newState = getBlockStateFromMeta(field_150939_a, meta);
         return placeBlockAt(stack, player, world, pos, facings[side],
-            hitX, hitY, hitZ, newState);
+                hitX, hitY, hitZ, newState);
     }
 
-	public boolean placeBlockAt(ItemStack stack, EntityPlayer player, World world, BlockPos pos,
-		EnumFacing face, float hitX, float hitY, float hitZ, IBlockState newState)
-    {
+    public boolean placeBlockAt(ItemStack stack, EntityPlayer player, World world, BlockPos pos,
+                                EnumFacing face, float hitX, float hitY, float hitZ, IBlockState newState) {
         int meta = getMetaFromBlockState(newState);
         return super.placeBlockAt(stack, player, world, pos.x, pos.y, pos.z, face.ordinal(),
-            hitX, hitY, hitZ, meta);
+                hitX, hitY, hitZ, meta);
     }
 
 }
